+++ CHANGELOG.md
@@ -5,3 +5,9 @@ added an `onChange` property to listen into 'set' calls
 ## 1.0.3 
 
 adding metadata to package.json for NPM
+
+## 1.0.4
+
+adding a type property for more semantic type checking
+executing onBadData in the context of the class
+allowing bad data to be set if onBadData returns true.